Add option to treat zero-velocity NoteOn as a NoteOn
This is against the specs, but some programs *still* export broken MIDIs with velocity 0 NoteOn events that are meant to be real NoteOn events, which breaks the visualizer.
This should be a configurable option though, as velocity 0 NoteOn is a NoteOff according to the official specs, so enabling this option may break some others.